Avoid re-setting the same ColumnSubstitutorPair in the hashmap
When vOld == 0, the CSP will have only null pointers, which are never used. When
vOld == the old v, then it will just overwrite the previous CSP set for v. This does not
happen when the hashmap is cleared, but doing so prevents the reuse of the CS and CSG
for different pixels in the same column.